+++ tools/cli/commands-packages.js
@@ -2005,7 +2005,8 @@ main.registerCommand({
     let changed = false;
 
     for (target of pluginsToAdd) {
-      let [id, version] = [target.substring(0,target.lastIndexOf('@')), target.substring(target.lastIndexOf('@')+1)];
+      let { id, version } =
+        require('../cordova/package-id-version-parser.js').parse(target);
       const newId = cordova.newPluginId(id);
 
       if (!(version && utils.isValidVersion(version, {forCordova: true}))) {